make db:build command more clear

+++ app/Console/Commands/MakeDatabaseCommand.php
@@ -13,7 +13,7 @@ class MakeDatabaseCommand extends Command
     public function handle(): void
     {
         if (in_array(config('app.env'), ['local', 'staging'])) {
-            if ($this->confirm('Do you wish to continue?')) {
+            if ($this->confirm('This will DELETE all data and re-migrate and seed. Do you wish to continue?')) {
                 $this->call('migrate:fresh');
                 $this->line('------');
                 $this->call('module:seed');
